+++ packages/terra-disclosure-manager/tests/jest/DisclosureManager.test.jsx
@@ -1,34 +1,36 @@
 import React from 'react';
-import DisclosureManager from '../../src/DisclosureManager';
+import DisclosureManager, { withDisclosureManager } from '../../src/DisclosureManager';
+
+const TestChild = withDisclosureManager(({ id }) => <div id={id} />);
 
 describe('DisclosureManager', () => {
   // Snapshot Tests
   it('should render using the provided render function', () => {
     const disclosureManager = (
-      <DisclosureManager
+      <DisclosureManager.WrappedComponent
         render={manager => (
           <div id="wrapper">
             {manager.children.components}
           </div>
         )}
       >
-        <div id="child1" />
-        <div id="child2" />
-      </DisclosureManager>
+        <TestChild id="child1" />
+        <TestChild id="child2" />
+      </DisclosureManager.WrappedComponent>
     );
 
-    const wrapper = shallow(disclosureManager);
+    const wrapper = mount(disclosureManager);
     expect(wrapper).toMatchSnapshot();
   });
 
-  const mountDisclosureManager = (diclosureTypes, renderFunc) => shallow((
-    <DisclosureManager
+  const mountDisclosureManager = (disclosureTypes, renderFunc) => mount((
+    <DisclosureManager.WrappedComponent
       render={renderFunc}
-      supportedDisclosureTypes={diclosureTypes}
+      supportedDisclosureTypes={disclosureTypes}
     >
-      <div id="child1" />
-      <div id="child2" />
-    </DisclosureManager>
+      <TestChild id="child1" />
+      <TestChild id="child2" />
+    </DisclosureManager.WrappedComponent>
   ));
 
   const validateInitialState = (wrapper) => {
@@ -41,24 +43,24 @@ describe('DisclosureManager', () => {
   };
 
   const validateChildAppDelegate = (wrapper) => {
-    const childApp1 = wrapper.find('#child1').getElements()[0].props.app;
-    expect(childApp1).toBeDefined();
-    expect(childApp1.disclose).toBeDefined();
+    const childDisclosureManager1 = wrapper.find('#child1').getElements()[1].props.disclosureManager;
+    expect(childDisclosureManager1).toBeDefined();
+    expect(childDisclosureManager1.disclose).toBeDefined();
 
-    const childApp2 = wrapper.find('#child2').getElements()[0].props.app;
-    expect(childApp2).toBeDefined();
-    expect(childApp2.disclose).toBeDefined();
+    const childDisclosureManager2 = wrapper.find('#child2').getElements()[1].props.disclosureManager;
+    expect(childDisclosureManager2).toBeDefined();
+    expect(childDisclosureManager2.disclose).toBeDefined();
   };
 
   const triggerChildDisclose = wrapper => (
     new Promise((resolve, reject) => {
-      const childApp1 = wrapper.find('#child1').getElements()[0].props.app;
-      childApp1.disclose({
+      const childDisclosureManager1 = wrapper.find('#child1').getElements()[1].props.disclosureManager;
+      childDisclosureManager1.disclose({
         preferredType: 'test',
         size: 'large',
         content: {
           key: 'DISCLOSE_KEY',
-          component: <div id="disclosure-component" />,
+          component: <TestChild id="disclosure-component" />,
         },
       }).then(resolve).catch(reject);
     })
@@ -73,17 +75,17 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
-        expect(disclosureContentApp).toBeDefined();
-        expect(disclosureContentApp.disclose).toBeDefined();
-        expect(disclosureContentApp.dismiss).toBeDefined();
-        expect(disclosureContentApp.closeDisclosure).toBeDefined();
-        expect(disclosureContentApp.goBack).toBeUndefined();
-        expect(disclosureContentApp.maximize).toBeDefined();
-        expect(disclosureContentApp.minimize).toBeUndefined();
-        expect(disclosureContentApp.requestFocus).toBeDefined();
-        expect(disclosureContentApp.releaseFocus).toBeUndefined();
-        expect(disclosureContentApp.registerDismissCheck).toBeDefined();
+        const disclosureContentManager = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
+        expect(disclosureContentManager).toBeDefined();
+        expect(disclosureContentManager.disclose).toBeDefined();
+        expect(disclosureContentManager.dismiss).toBeDefined();
+        expect(disclosureContentManager.closeDisclosure).toBeDefined();
+        expect(disclosureContentManager.goBack).toBeUndefined();
+        expect(disclosureContentManager.maximize).toBeDefined();
+        expect(disclosureContentManager.minimize).toBeUndefined();
+        expect(disclosureContentManager.requestFocus).toBeDefined();
+        expect(disclosureContentManager.releaseFocus).toBeUndefined();
+        expect(disclosureContentManager.registerDismissCheck).toBeDefined();
       })
   );
 
@@ -117,13 +119,13 @@ describe('DisclosureManager', () => {
     validateChildAppDelegate(wrapper);
 
     return new Promise((resolve, reject) => {
-      const childApp1 = wrapper.find('#child1').getElements()[0].props.app;
+      const childApp1 = wrapper.find('#child1').getElements()[1].props.disclosureManager;
       childApp1.disclose({
         preferredType: 'test',
         size: 'large',
         content: {
           key: 'DISCLOSE_KEY',
-          component: <div id="disclosure-component" />,
+          component: <TestChild id="disclosure-component" />,
         },
       }).then(({ dismissDisclosure, afterDismiss }) => {
         dismissDisclosureInstance = dismissDisclosure;
@@ -169,7 +171,7 @@ describe('DisclosureManager', () => {
 
     return triggerChildDisclose(wrapper)
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.maximize().then(resolve).catch(reject);
       }))
@@ -184,7 +186,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -197,7 +199,7 @@ describe('DisclosureManager', () => {
         expect(disclosureContentApp.registerDismissCheck).toBeDefined();
       })
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.minimize().then(resolve).catch(reject);
       }))
@@ -212,7 +214,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -238,13 +240,13 @@ describe('DisclosureManager', () => {
     validateChildAppDelegate(wrapper);
 
     return new Promise((resolve, reject) => {
-      const childApp1 = wrapper.find('#child1').getElements()[0].props.app;
+      const childApp1 = wrapper.find('#child1').getElements()[1].props.disclosureManager;
       childApp1.disclose({
         preferredType: 'test',
         size: 'fullscreen',
         content: {
           key: 'DISCLOSE_KEY',
-          component: <div id="disclosure-component" />,
+          component: <TestChild id="disclosure-component" />,
         },
       }).then(resolve).catch(reject);
     })
@@ -259,7 +261,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -286,7 +288,7 @@ describe('DisclosureManager', () => {
 
     return triggerChildDisclose(wrapper)
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.requestFocus().then(resolve).catch(reject);
       }))
@@ -301,7 +303,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -314,7 +316,7 @@ describe('DisclosureManager', () => {
         expect(disclosureContentApp.registerDismissCheck).toBeDefined();
       })
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.releaseFocus().then(resolve).catch(reject);
       }))
@@ -329,7 +331,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.key).toEqual('DISCLOSE_KEY');
         expect(wrapper.state().disclosureComponentData.DISCLOSE_KEY.component.props.id).toEqual('disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -356,7 +358,7 @@ describe('DisclosureManager', () => {
 
     return triggerChildDisclose(wrapper)
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.dismiss().then(resolve).catch(reject);
       }))
@@ -385,7 +387,7 @@ describe('DisclosureManager', () => {
 
     return triggerChildDisclose(wrapper)
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.closeDisclosure().then(resolve).catch(reject);
       }))
@@ -414,14 +416,14 @@ describe('DisclosureManager', () => {
 
     return triggerChildDisclose(wrapper)
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.disclose({
           preferredType: 'test',
           size: 'huge',
           content: {
             key: 'NESTED',
-            component: <div id="nested-disclosure-component" />,
+            component: <TestChild id="nested-disclosure-component" />,
           },
         }).then(resolve).catch(reject);
       }))
@@ -436,7 +438,7 @@ describe('DisclosureManager', () => {
         expect(wrapper.state().disclosureComponentData.NESTED.key).toEqual('NESTED');
         expect(wrapper.state().disclosureComponentData.NESTED.component.props.id).toEqual('nested-disclosure-component');
 
-        const disclosureContentApp = wrapper.find('#nested-disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#nested-disclosure-component').getElements()[1].props.disclosureManager;
         expect(disclosureContentApp).toBeDefined();
         expect(disclosureContentApp.disclose).toBeDefined();
         expect(disclosureContentApp.dismiss).toBeDefined();
@@ -449,7 +451,7 @@ describe('DisclosureManager', () => {
         expect(disclosureContentApp.registerDismissCheck).toBeDefined();
       })
       .then(() => new Promise((resolve, reject) => {
-        const disclosureContentApp = wrapper.find('#nested-disclosure-component').getElements()[0].props.app;
+        const disclosureContentApp = wrapper.find('#nested-disclosure-component').getElements()[1].props.disclosureManager;
 
         disclosureContentApp.goBack().then(resolve).catch(reject);
       }))
@@ -495,7 +497,7 @@ describe('DisclosureManager', () => {
         expect(firstRenderPayload.dismissPresentedComponent).toBeDefined();
         expect(firstRenderPayload.closeDisclosure).toBeDefined();
         expect(firstRenderPayload.children).toBeDefined();
-        expect(firstRenderPayload.children.components.length).toEqual(2);
+        expect(firstRenderPayload.children.components.props.children.length).toEqual(2);
 
         expect(firstRenderPayload.disclosure).toBeDefined();
         expect(firstRenderPayload.disclosure.isOpen).toBeFalsy();
@@ -508,7 +510,7 @@ describe('DisclosureManager', () => {
         expect(secondRenderPayload.dismissPresentedComponent).toBeDefined();
         expect(secondRenderPayload.closeDisclosure).toBeDefined();
         expect(secondRenderPayload.children).toBeDefined();
-        expect(secondRenderPayload.children.components.length).toEqual(2);
+        expect(secondRenderPayload.children.components.props.children.length).toEqual(2);
 
         expect(secondRenderPayload.disclosure).toBeDefined();
         expect(secondRenderPayload.disclosure.isOpen).toBeTruthy();